Button to Display Map
Button labled Floor Map will display map of HackRU. Clicking on it will save the position in terms of x and y coordinates in the program and will allow the user to confirm.

Things to be fixed/added: Sometimes the Help me! button doesn't become more saturated but is still clickable after dealing with the map UI. Also, I cannot get the pin image to show up. Finally, the locations need to be stored somehow (array?) such that both x and y positions are relayed.